Move functions out of mob.h and cleanup
